+++ src/settings/descriptors/staleDataDescriptor.ts
@@ -30,6 +30,7 @@ export class StaleDataDescriptor extends BaseDescriptor {
     public showLatterAvailableValue: boolean = false;
     public staleDataText: string = undefined;
     public staleDataThreshold: number = 0;
+    public deductThresholdDays: boolean = false;
     public color: string = "#3599b8";
     public background: string = "";
 }