+++ src/visualComponent/rootComponent.ts
@@ -212,6 +212,7 @@ export class RootComponent extends BaseContainerComponent<IVisualComponentConstr
         this.subtitleComponent.render({
             dateDifference: data.dateDifference,
             series: data.series,
+            staleDataDifference: data.staleDateDifference,
             staleDataSettings: settings.staleData,
             subtitleSettings: settings.subtitle,
             warningState: data.warningState,